requirements.txt
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-filter==2.2.0
django-import-export==2.0.1
django-phone-verify==1.0.0
django-phonenumber-field==4.0.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
phonenumbers==8.11.2
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.5
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
twilio==6.35.2
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
